ductm commit

+++ API/CinemaManagement/DTOs/CmsDtos/ShowTime/AddShowTime.cs
@@ -7,7 +7,7 @@ namespace CinemaManagement.DTOs.CmsDtos.ShowTime
     public class AddShowTime : Entity<long?>
     {
         public DateTime StartTime { get; set; }
-        public int MovieId { get; set; }
-        public int RoomId { get; set; }
+        public long MovieId { get; set; }
+        public long RoomId { get; set; }
     }
 }
